requirements.txt
adblockparser==0.7
AdvancedHTTPServer==2.2.0
aiocmd==0.1.2
aioconsole==0.3.1
aiodns==3.0.0
aiofiles==0.8.0
aiohttp==3.8.1
aiomultiprocess==0.8.0
aioredis==1.3.1
aiosignal==1.2.0
aiosmb==0.2.50
aiosqlite==0.17.0
aiowinreg==0.0.7
ajpy==0.0.4
alembic==1.7.6.dev0
altgraph==0.17.2
aniso8601==9.0.1
anyio==3.6.1
apispec==5.2.2
apispec-webframeworks==0.5.2
appdirs==1.4.4
APScheduler==3.9.1
arrow==1.2.2
asciitree==0.3.3
asgiref==3.5.0
asn1crypto==1.5.1
async-timeout==4.0.1
asysocks==0.1.2
attrs==21.2.0
autobahn==22.1.1
Automat==20.2.0
autopep8==2.0.0
Babel==2.8.0
backcall==0.2.0
backoff==1.11.1
base58==1.0.3
bcrypt==4.0.1
beautifulsoup4==4.11.1
bidict==0.21.3
binwalk==2.3.3
bleach==5.0.0
blinker==1.4
bluepy==1.3.0
boltons==21.0.0
bottle==0.12.21
Bottleneck==1.3.2
Brlapi==0.8.4
Brotli==1.0.9
catfish==4.16.4
cbor==1.0.0
censys==2.0.9
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.6
cheroot==8.6.0+ds1
CherryPy==18.8.0
cherrypy-cors==1.6
cli-helpers==2.2.1
click==8.1.3
click-plugins==1.1.1
cmd2==0.0.0
colorama==0.4.6
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
crackmapexec==5.2.2
cryptography==3.4.8
cycler==0.11.0
Cython==0.29.30
dataclasses-json==0.5.7
dbus-python==1.2.18
debtags==2.1
decorator==4.4.2
defusedxml==0.7.1
dicttoxml==1.7.4
distlib==0.3.5
distro==1.7.0
Django==3.2.13
dnslib==0.9.19
dnspython==2.2.1
docopt==0.6.2
donut-shellcode==0.9.3
dropbox==11.30.0
dsinternals==1.2.4
ecdsa==0.18.0
email-validator==1.1.2
et-xmlfile==1.0.1
ExifRead==3.0.0
faraday-agent-parameters-types==1.0.3
faraday-cli==2.1.6
faraday-plugins==1.6.8
faradaysec==4.0.4
fastapi==0.74.1
feedparser==6.0.8
fierce==1.5.0
filedepot==0.5.2
filelock==3.7.1
filteralchemy==0.1.0
flasgger==0.9.5
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Classful==0.14.1
Flask-Cors==3.0.10
Flask-KVSession-fork==0.6.4
Flask-Limiter==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==4.0.0
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==3.0.2
Flask-WTF==0.14.3
flatbuffers==2.0.6+dfsg1.1
fonttools==4.32.0
frozenlist==1.2.0
fs==2.4.16
future==0.18.2
GDAL==3.5.1
geoip2==2.9.0
geojson==2.5.0
gitdb==4.0.9
GitPython==3.1.27
gpg==1.17.1
graphene==2.1.9
graphene-sqlalchemy==2.1.2
graphql-core==2.2.1
graphql-relay==2.0.1
greenlet==2.0.1
h11==0.13.0
h2==4.1.0
hashID==3.1.4
hiredis==1.0.1
hpack==4.0.0
html2text==2020.1.16
html5lib==1.1
httpagentparser==1.9.1
httpcore==0.15.0
httpx==0.23.0
humanize==0.0.0
hupper==1.10.3
hyperframe==6.0.0
hyperlink==21.0.0
icalendar==4.0.3
idna==3.3
impacket==0.10.0
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==1.1.1
invoke==1.7.0
ipwhois==0.15.1
IPy==1.1
ipython==7.31.1
itsdangerous==2.1.2
jaraco.classes==3.2.1
jaraco.collections==3.5.1
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.7.0
jdcal==1.0
jedi==0.18.0
Jinja2==3.1.2
jq==1.2.1
jsonschema==3.2.0
kaitaistruct==0.9
kali-tweaks==2022.3.0
KismetCaptureBtGeiger==2021.7.1
KismetCaptureFreaklabsZigbee==2018.7.0
KismetCaptureRtl433==2020.10.1
KismetCaptureRtladsb==2020.10.1
KismetCaptureRtlamr==2020.10.1
kiwisolver==1.3.2
ldap3==2.9.1
ldapdomaindump==0.9.3
lightdm-gtk-greeter-settings==1.2.2
limiter==0.1.2
limits==1.5.1
llvmlite==0.38.1
log-symbols==0.0.14
louis==3.22.0
lsassy==3.1.1
lxml==4.9.1
lz4==4.0.0+dfsg
macholib==1.16
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.28.0
matplotlib==3.5.2
matplotlib-inline==0.1.3
maxminddb==2.0.3
mechanize==0.4.8
minidump==0.0.21
minikerberos==0.2.14
mistune0==0.8.4
mitmproxy==8.1.1
mnemonic==0.19
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.3
msldap==0.3.30
multidict==5.1.0
mypy-extensions==0.4.3
mysqlclient==1.4.6
nassl==4.0.2
neo4j==1.7.0.dev0
neobolt==1.7.17
neotime==1.7.4
netaddr==0.8.0
netifaces==0.11.0
networkx==2.6.3
nplusone==1.0.0
ntlm-auth==1.4.0
numba==0.55.2
numexpr==2.8.3
numpy==1.21.5
odfpy==1.4.2
olefile==0.46
onboard==1.4.1
openpyxl==3.0.9
packaging==21.3
pandas==1.3.5
paramiko==2.10.4
parso==0.8.1
passlib==1.7.4
Paste==3.5.0
PasteDeploy==2.1.1
PasteScript==2.0.2
patator==0.9
pcapy==0.11.5.dev0
pefile==2022.5.30
pendulum==2.1.2
pexpect==4.8.0
pgcli==3.4.1
pgspecial==1.11.10
phonenumbers==8.12.1
pickleshare==0.7.5
Pillow==9.2.0
plaster==1.0
plaster-pastedeploy==0.5
platformdirs==2.5.2
pluggy==1.0.0+repack
pluginbase==1.0.1
ply==3.11
portend==3.1.0
prettytable==2.5.0
promise==2.3
prompt-toolkit==3.0.30
protobuf==4.21.10
psycopg2==2.9.2
ptyprocess==0.7.0
publicsuffix2==2.20191221
publicsuffixlist==0.7.10
py==1.10.0
py-sneakers==1.0.1
py-ubjson==0.16.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycares==4.1.2
pycodestyle==2.9.1
pycparser==2.21
pycryptodomex==3.11.0
pycurl==7.44.1
pydantic==1.9.0
PyDispatcher==2.0.5
pydot==1.4.2
pyee==9.0.4
pyExploitDb==0.2.0
pyfiglet==0.8.post0
pygame==2.1.2
pygexf==0.2.2
Pygments==2.12.0
PyGObject==3.42.1
pygraphviz==1.7
PyHamcrest==2.0.2
pyinotify==0.9.6
PyInstaller==3.5+498e6ee058
PyJWT==2.4.0
pylnk3==0.4.2
pyminifier==2.1
pymssql==2.2.2
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyotp==2.6.0
pyparsing==3.0.7
PyPDF2==2.6.0
pyperclip==1.8.2
pypng==0.0.20
pyppeteer==0.2.5
pypsrp==0.5.0
pypykatz==0.4.9
PyQRCode==1.2.1
PyQt5==5.15.7
PyQt5-sip==12.11.0
pyqtgraph==0.12.4
pyramid==2.0
pyrsistent==0.18.1
pyserial==3.5
pyShodan==0.2.3
pysmi==0.3.2
pysnmp==4.4.12
PySocks==1.7.1
pyspnego==0.1.5
pytest==7.1.2
python-apt==2.3.0+b2
python-dateutil==2.8.1
python-debian==0.1.46
python-docx==0.8.11
python-dotenv==0.20.0
python-engineio==4.0.0
python-magic==0.4.26
python-pam==2.0.2
python-pptx==0.6.18
python-slugify==4.0.0
python-snappy==0.5.3
python-socketio==5.0.3
python-status==1.0.1
PyTrie==0.4.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
pyVNC==0.1
pywerview==0.3.2
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.3.1
Quamash==0.6.1
redis==3.5.3
repoze.lru==0.7
requests==2.27.1
requests-file==1.5.1
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
responses==0.18.0
retrying==1.3.3
rfc3986==1.5.0
rich==12.4.4
Routes==2.5.1
rq==1.10.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
rule-engine==1.1.0
Rx==3.2.0
scapy==2.4.4
scipy==1.7.3
secure==0.3.0
service-identity==18.1.0
setproctitle==1.2.2
sgmllib3k==1.0.0
shodan==1.28.0
simple-rest-client==1.1.3
simplejson==3.17.6
simplekv==0.13.0
six==1.16.0
slowapi==0.1.4
smmap==5.0.0
smoke-zephyr==2.0.1
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2
speaklater==1.4
spinners==0.0.24
spyse-python==2.2.3
SQLAlchemy==1.4.44
sqlalchemy-schemadisplay==1.3
SQLAlchemy-Utc==0.11.0
sqlparse==0.4.2
sslyze==5.0.5
starlette==0.18.0
stone==3.3.1
sympy==1.10.1
syslog-rfc5424-formatter==1.2.2
tables==3.7.0
tabulate==0.8.9
Tempita==0.5.2
tempora==5.0.1
termcolor==1.1.0
terminaltables==3.1.0
texttable==1.6.4
theHarvester==4.0.3
tld==0.11.11
tls-parser==1.2.2
token-bucket==0.3.0
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
translationstring==1.4
Twisted==22.4.0
txaio==21.2.1
typing-extensions==3.10.0.2
typing-inspect==0.7.1
tzlocal==4.2
u-msgpack-python==2.3.0
ufoLib2==0.13.1
ujson==5.4.0
unicodecsv==0.14.1
unicodedata2==14.0.0
Unidecode==1.3.4
urllib3==1.26.9
urwid==2.1.2
uvicorn==0.17.6
uvloop==0.16.0
validators==0.20.0
venusian==3.0.0
virtualenv==20.15.0+ds
wafw00f==2.2.0
wapiti3==3.0.4
wcwidth==0.2.5
webargs==8.0.1
webencodings==0.5.1
WebOb==1.8.6
websocket-client==1.2.3
websockets==10.2
websockify==0.10.0
Werkzeug==2.2.2
wfuzz==3.1.0
whois==0.8
wifite==2.6.0
winacl==0.1.2
wsaccel==0.6.3
wsproto==1.1.0
WTForms==2.2.1
xdg==5
xlrd==1.2.0
XlsxWriter==3.0.2
xlutils==2.0.0
xlwt==1.3.0
xmltodict==0.12.0
yara-python==4.2.0
yarl==1.7.2
yaswfp==0.9.3
zc.lockfile==2.0
zipp==1.0.0
zlib-wrapper==0.1.3
zope.deprecation==4.4.0
zope.interface==5.4.0
